refactor: ui 디렉토리 정리 - roominfo, util 이동 #1
