+++ app/src/main/java/com/ganaljigi/kubf/feature/home/component/HomeToggle.kt
@@ -21,7 +21,7 @@ import com.ganaljigi.kubf.feature.home.viewmodel.ToggleUiState
 import com.ganaljigi.kubf.core.designsystem.theme.MainGreen
 import com.ganaljigi.kubf.core.designsystem.theme.KUBFAndroidTheme
 import com.ganaljigi.kubf.core.designsystem.theme.LightGreen
-import com.ganaljigi.kubf.ui.util.noRippleClickable
+import com.ganaljigi.kubf.core.ui.util.noRippleClickable
 
 @Composable
 fun HomeToggle(
