+++ app/src/main/java/com/ganaljigi/kubf/feature/home/component/bottomsheet/HomeSearchBottomSheetEmptyResult.kt
@@ -20,7 +20,7 @@ import com.ganaljigi.kubf.core.designsystem.theme.Gray4
 import com.ganaljigi.kubf.core.designsystem.theme.KUBFAndroidTheme
 import com.ganaljigi.kubf.core.designsystem.theme.LightGreen
 import com.ganaljigi.kubf.core.designsystem.theme.MainGreen
-import com.ganaljigi.kubf.ui.util.noRippleClickable
+import com.ganaljigi.kubf.core.ui.util.noRippleClickable
 
 @Composable
 fun HomeSearchBottomSheetEmptyResult(
