+++ app/src/main/java/com/ganaljigi/kubf/feature/home/component/bottomsheet/HomeSearchBottomSheetWithItemList.kt
@@ -28,8 +28,8 @@ import com.ganaljigi.kubf.core.designsystem.theme.Gray1
 import com.ganaljigi.kubf.core.designsystem.theme.Gray3
 import com.ganaljigi.kubf.core.designsystem.theme.KUBFAndroidTheme
 import com.ganaljigi.kubf.core.designsystem.theme.MainGreen
-import com.ganaljigi.kubf.ui.util.noRippleClickable
-import com.ganaljigi.kubf.ui.util.noRippleClickableSingle
+import com.ganaljigi.kubf.core.ui.util.noRippleClickable
+import com.ganaljigi.kubf.core.ui.util.noRippleClickableSingle
 
 @Composable
 fun HomeSearchBottomSheetWithItemList(
