+++ app/src/main/java/com/ganaljigi/kubf/feature/home/component/bottomsheet/HomeSearchBottomSheetWithSingleItem.kt
@@ -23,7 +23,7 @@ import com.ganaljigi.kubf.feature.home.model.SearchResult
 import com.ganaljigi.kubf.core.designsystem.theme.Gray3
 import com.ganaljigi.kubf.core.designsystem.theme.KUBFAndroidTheme
 import com.ganaljigi.kubf.core.designsystem.theme.MainGreen
-import com.ganaljigi.kubf.ui.util.noRippleClickable
+import com.ganaljigi.kubf.core.ui.util.noRippleClickable
 
 @Composable
 fun HomeSearchBottomSheetSingleItem(
