+++ app/src/main/java/com/ganaljigi/kubf/feature/home/component/map/MarkerComposableInfoWindow.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.painter.Painter
 import com.ganaljigi.kubf.feature.home.viewmodel.SpecialMarkerInfo
-import com.ganaljigi.kubf.ui.util.noRippleClickable
+import com.ganaljigi.kubf.core.ui.util.noRippleClickable
 import com.google.maps.android.compose.MarkerInfoWindow
 import com.google.maps.android.compose.MarkerState
