+++ app/src/main/java/com/ganaljigi/kubf/feature/home/component/search/HomeSearchBar.kt
@@ -26,7 +26,7 @@ import com.ganaljigi.kubf.core.designsystem.component.KUBFSearchBar
 import com.ganaljigi.kubf.core.designsystem.theme.Gray1
 import com.ganaljigi.kubf.core.designsystem.theme.KUBFAndroidTheme
 import com.ganaljigi.kubf.core.designsystem.theme.MainGreen
-import com.ganaljigi.kubf.ui.util.noRippleClickable
+import com.ganaljigi.kubf.core.ui.util.noRippleClickable
 
 @Composable
 fun HomeSearchBar(
