+++ app/src/main/java/com/ganaljigi/kubf/feature/home/component/search/HomeSearchContent.kt
@@ -27,9 +27,9 @@ import com.ganaljigi.kubf.core.designsystem.theme.Gray3
 import com.ganaljigi.kubf.core.designsystem.theme.Gray4
 import com.ganaljigi.kubf.core.designsystem.theme.KUBFAndroidTheme
 import com.ganaljigi.kubf.core.designsystem.theme.MainGreen
-import com.ganaljigi.kubf.ui.util.noRippleClickable
-import com.ganaljigi.kubf.ui.util.noRippleClickableSingle
-import com.ganaljigi.kubf.ui.util.toAnnotatedString
+import com.ganaljigi.kubf.core.ui.util.noRippleClickable
+import com.ganaljigi.kubf.core.ui.util.noRippleClickableSingle
+import com.ganaljigi.kubf.core.ui.util.toAnnotatedString
 
 @Composable
 fun HomeSearchContent(
