+++ app/src/main/java/com/ganaljigi/kubf/feature/room/component/RoomPicComponent.kt
@@ -1,5 +1,5 @@
 @file:OptIn(androidx.compose.foundation.ExperimentalFoundationApi::class)
-package com.ganaljigi.kubf.ui.roominfo.component
+package com.ganaljigi.kubf.feature.room.component
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
